some aesthetic changes

+++ dump_mhex_to_mat.m
@@ -17,12 +17,12 @@ function dump_mhex_to_mat(G, filename)
 % the first matrix multiplication, turning num_v raw category scores into
 % num_leaf raw assignment scores
 % matrix size: num_leaf * num_v
-M1 = single(G.S);
+M1 = double(G.S);
 
 % the second matrix multiplication, turning num_leaf assignment
 % probabilities into num_v marginal category probabilities
 % matrix size: num_v * num_leaf
-M2 = single(G.E_des_i(:, G.leaves));
+M2 = double(G.E_des_i(:, G.leaves));
 
 save(filename, 'M1', 'M2', '-v7.3');
 fprintf('successfully dumped MHEX Graph to %s\n', filename);
